ColoredObject is complete as far as I know.
Ball needs the path functionality, which cannot be done until Parabola is done.

Parabola exists but is empty.

GameGUI just has a black Ball going back and forth as a test. It's not the smoothest thing in the world, but it works

CircleFragment, Circle, and GameLogic all have yet to be created