CircleFragment class was started
CircleFragment only has a constructor and an incomplete draw method

GameGUI was changed from an Applet to a JFrame, which makes the animation smoother but more flickery. It also should make adding the key listeners and building the final version an easier process.

Removed print statements in Parabola that were for debugging purposes.

Verified that the equals method in ColoredObject works properly